requirements.txt
requests~=2.28.1
pyfiglet~=0.8.post1
python-nmap~=0.7.1